add refined security on user

+++ user/data_user.php
@@ -1,17 +1,6 @@
 <?php
-session_start();
 
-if(!isset($_SESSION['username'])){
-    header("Location: ../index.php");
-    exit;
-}
-
-/* cek role admin */
-if($_SESSION['role'] != 'admin'){
-    echo "Akses ditolak!";
-    exit;
-}
-
+include("../config/auth.php");
 include("../config/koneksi.php");
 
 /* ambil data user */
@@ -61,8 +50,10 @@ $query = mysqli_query($koneksi,"SELECT * FROM user");
 
         <div class="table-container">
 
+        <?php if($isAdmin){ ?>
             <a href="tambah_user.php" class="btn-tambah">+ Tambah User</a>
-
+        <?php } ?>
+        
             <table>
 
                 <thead>
@@ -92,8 +83,10 @@ $query = mysqli_query($koneksi,"SELECT * FROM user");
                         <td><?php echo $row['username']; ?></td>
                         <td><?php echo $row['role']; ?></td>
                         <td>
-                        <a href="edit.php?id=<?php echo $row['id_user']; ?>" class="btn-edit">Edit</a>
-                        <a href="hapus_user.php?id=<?php echo $row['id_user']; ?>"class="btn-hapus"onclick="return konfirmasiHapus('Yakin ingin menghapus data ini?')">Hapus</a>
+                            <?php if($isAdmin){ ?>
+                                <a href="edit.php?id=<?php echo $row['id_user']; ?>" class="btn-edit">Edit</a>
+                                <a href="hapus_user.php?id=<?php echo $row['id_user']; ?>"class="btn-hapus"onclick="return konfirmasiHapus('Yakin ingin menghapus data ini?')">Hapus</a>
+                            <?php } ?>
                         </td>
 
                     </tr>
